Improve install

+++ vim/install.sh
@@ -3,7 +3,7 @@
 set -o errexit
 set -o nounset
 
-readonly CURRENT_DIR=$(dirname "$(readlink -f "$0")")
+readonly DIR="$( cd "$( dirname "$0" )" && pwd )"
 
 check_version() {
   echo "== VIM/NeoVim configuration"
@@ -29,25 +29,27 @@ symlink_config() {
   readonly TARGET_NVIM="$HOME/.config/nvim"
 
   mkdir -p "$HOME/.config/"
-  echo "=== Installing vim/nvim configuration (vimrc and vim folder from $CURRENT_DIR)"
+  echo "=== Installing vim/nvim configuration (vimrc and vim folder from $DIR)"
   mv "${TARGET_VIM}rc" "$TARGET_VIM" "$HOME/.ideavimrc" "$TARGET_NVIM" /tmp/ || true
 
-  ln -sf "$CURRENT_DIR" "$TARGET_VIM"
+  ln -sf "$DIR" "$TARGET_VIM"
   ln -sf "$TARGET_VIM/init.vim" "${TARGET_VIM}rc"
-  ln -sf "$CURRENT_DIR/ideavimrc" "$HOME/.ideavimrc"
+  ln -sf "$DIR/ideavimrc" "$HOME/.ideavimrc"
 
-  ln -sf "$CURRENT_DIR" "$TARGET_NVIM"
+  ln -sf "$DIR" "$TARGET_NVIM"
 }
 
 install_plugins() {
   echo "=== Installing vim plugins..."
-  curl -fLo ~/.vim/autoload/plug.vim --create-dirs \
-      https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
-  rm ~/.local/share/nvim/site/autoload/plug.vim || true
-  ln -sf ~/.vim/autoload/plug.vim ~/.local/share/nvim/site/autoload/plug.vim
+  if [ ! -s ~/.local/share/nvim/site/autoload/plug.vim ]; then
+    curl -fLo ~/.vim/autoload/plug.vim --create-dirs \
+        https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
+    rm ~/.local/share/nvim/site/autoload/plug.vim || true
+    ln -sf ~/.vim/autoload/plug.vim ~/.local/share/nvim/site/autoload/plug.vim
+  fi
   time \
     $VIMCMD -u ~/.vim/01_plugins.vim \
-    -c '+verbose' '+PlugUpdate' '+PlugClean' '+UpdateRemotePlugins' '+messages' +qall \
+    -c '+verbose' '+PlugUpgrade' '+PlugUpdate' '+PlugClean' '+UpdateRemotePlugins' '+CheckHealth' '+messages' +qall \
     foo_file || true
   # https://stackoverflow.com/questions/12213597/how-to-see-which-plugins-are-making-vim-slow
   $VIMCMD$VIMFLAGS --cmd 'profile start profile.log' \
@@ -58,10 +60,25 @@ install_plugins() {
     -c 'qa!'
 }
 
+install_extensions() {
+  # https://github.com/roxma/nvim-completion-manager#installation
+  if hash pip3 2> /dev/null; then
+    pip3 install --user neovim jedi mistune psutil setproctitle websocket-client sexpdata
+  else
+    "pip3 not available, skipping python3 nvim extensions"
+  fi
+  if hash pip 2> /dev/null; then
+    pip install --user neovim jedi websocket-client sexpdata
+  else
+    "pip not available, skipping python nvim extensions"
+  fi
+}
+
 main() {
   check_version
   symlink_config
+  install_extensions
   install_plugins
 }
 
-main "$@"
+time main "$@"
